Admin Payment Dashboard

+++ app/src/main/java/com/example/food_ordering_app/AdminUserActivity.java
@@ -74,6 +74,11 @@ public class AdminUserActivity extends AppCompatActivity {
                     startActivity(intent);
                     return true;
                 }
+                if(id==R.id.admin_payment){
+                    intent = new Intent(AdminUserActivity.this, AdminPaymentActivity.class);
+                    startActivity(intent);
+                    return true;
+                }
                 if (id == R.id.logout) {
                     SharedPreferences.Editor editor = sharedPreferences.edit();
                     editor.clear();
